+++ src/replace.ts
@@ -1,0 +1,6 @@
+export function replace(
+  node: ChildNode,
+  ...newNodes: ReadonlyArray<string | Node>
+): void {
+  node.replaceWith(...newNodes);
+}